monitor/pyproject.toml
[tool.poetry]
name = "monitor"
version = "0.1.0"
description = ""
authors = ["Sway Ducky <swayducky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.3.10"
pyobjc-framework-quartz = "^9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
